% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pie_plot.R
\name{pie_plot}
\alias{pie_plot}
\title{Simple Pie Plot}
\usage{
pie_plot(
  df,
  var_name,
  main_title = NULL,
  legend_title = NULL,
  label_size = NULL,
  palette = NULL
)
}
\arguments{
\item{df}{Data frame containing factor variable to plot.}

\item{var_name}{Factor variable name.}

\item{main_title}{Optional.}

\item{legend_title}{Optional.}

\item{label_size}{Optional, numeric. Size of the text label.}

\item{palette}{Optional. See R Color Brewer’s palettes.}
}
\value{
A simple pie chart where the labels correspond to the percentage of each category.
}
\description{
Simple Pie Plot
}
\examples{
my_data=data.frame(category = c("A", "B", "A", "C", "B", "B", "C", "A"))
pie_plot(my_data,"category")
}
